Little fix

+++ src/storage/txn/commands/acquire_pessimistic_lock.rs
@@ -176,7 +176,7 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for AcquirePessimisticLock 
 
         let rows = if res.is_ok() { total_keys } else { 0 };
 
-        record_network_out_bytes(res.as_ref().map_or(0, |v| v.estimate_resp_size() as u64));
+        record_network_out_bytes(res.as_ref().map_or(0, |v| v.estimate_resp_size()));
         let pr = ProcessResult::PessimisticLockRes { res };
 
         let to_be_write = make_write_data(modifies, old_values);
